mon/OSDMonitor: Add a safety builtin to mon_osd_report_timeout
Add a safety builtin that when osd timeout happens, it will check the
relationship between mon_osd_report_timout and osd_beacon_report_interval
and make sure the mon_osd_report_timout to be set a value that is 2x what is
set for osd_beacon_report_interval. This will allow for the OSD to miss a
beacon and catch it on the next beacon without being marked down by the monitor.

Fixes: https://tracker.ceph.com/issues/40668
Signed-off-by: luo rixin <luorixin@huawei.com>